requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-jose[cryptography]
python-dotenv
aiogram
passlib[bcrypt]